requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64


absl-py=2.1.0=pyhd8ed1ab_0
appnope=0.1.2=py37hf985489_1
argcomplete=1.12.3=pyhd8ed1ab_2
astor=0.8.1=pyh9f0ad1d_0
autograd=1.3=py_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=2.120=openblas
blas-devel=3.9.0=20_osx64_openblas
brotli=1.1.0=h0dc2134_1
brotli-bin=1.1.0=h0dc2134_1
brotli-python=1.0.9=py37h0582d14_7
bzip2=1.0.8=h10d778d_5
c-ares=1.25.0=h10d778d_0
ca-certificates=2023.11.17=h8857fd0_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.16.0=he43a7df_1008
cctools_osx-64=973.0.1=hcc6d90d_12
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
clang=12.0.1=h694c41f_4
clang-12=12.0.1=default_he082bbe_4
clang_osx-64=12.0.1=h633439f_9
clangxx=12.0.1=default_he082bbe_4
clangxx_osx-64=12.0.1=hdb584c0_9
colorama=0.4.6=pyhd8ed1ab_0
compiler-rt=12.0.1=he01351e_0
compiler-rt_osx-64=12.0.1=hd3f61c9_0
cvxopt=1.3.0=py37h86ab292_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h811a1a6_3
debugpy=1.4.1=py37hd8d24ac_0
decorator=5.1.0=pyhd8ed1ab_0
dsdp=5.8=h6e329d1_1203
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.5.0=hf0c8a7f_1
ffmpeg=4.3.2=hbf27d7b_3
fftw=3.3.10=nompi_h4fa670e_108
fontconfig=2.14.2=h5bb23bf_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h60636b9_2
future=0.18.2=py37hf985489_5
gast=0.5.4=pyhd8ed1ab_0
gettext=0.21.1=h8a4c099_0
glpk=4.65=h0f52abe_1004
gmp=6.3.0=h93d8f39_0
gmpy2=2.1.2=py37h60f582e_0
gnutls=3.6.13=h756fd2b_1
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.13=h2e338ed_1001
grpcio=1.41.1=py37hac9a2c0_0
gsl=2.7=h93259b0_0
h5py=3.4.0=nompi_py37h0ac0de7_101
harfbuzz=2.9.1=h159f659_1
hdf5=1.12.1=nompi_h0aa1fa2_104
icu=68.2=he49afe7_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=4.8.1=py37hf985489_0
importlib_metadata=4.8.1=hd8ed1ab_0
ipykernel=6.4.2=py37h4c52d7d_0
ipython=7.28.0=py37h4c52d7d_0
ipython_genutils=0.2.0=py_1
jasper=1.900.1=h636a363_1006
jbig=2.1=h0d85af4_2003
jedi=0.18.0=py37hf985489_2
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=hb7f2c08_3
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.8.1=py37hf985489_0
keras=2.3.1=py37_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py37h737db71_0
krb5=1.19.3=hb98e516_0
lame=3.100=hb7f2c08_1003
lcms2=2.12=h577c468_0
ld64_osx-64=609=hc6b1d1e_10
lerc=3.0=he49afe7_0
libblas=3.9.0=20_osx64_openblas
libbrotlicommon=1.1.0=h0dc2134_1
libbrotlidec=1.1.0=h0dc2134_1
libbrotlienc=1.1.0=h0dc2134_1
libcblas=3.9.0=20_osx64_openblas
libclang=11.1.0=default_he082bbe_1
libclang-cpp12=12.0.1=default_he082bbe_4
libcurl=7.81.0=h97da3c1_0
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.8=h0d85af4_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=h10d778d_2
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_2
libgfortran5=13.2.0=h2873a65_2
libglib=2.78.3=h198397b_0
libgpuarray=0.7.6=h0d85af4_1003
libiconv=1.17=hd75f5a5_2
liblapack=3.9.0=20_osx64_openblas
liblapacke=3.9.0=20_osx64_openblas
libllvm11=11.1.0=h8fb7429_5
libllvm12=12.0.1=hd011deb_2
libllvm14=14.0.6=hc8e404f_4
libnghttp2=1.58.0=h64cf6d3_1
libopenblas=0.3.25=openmp_hfef2a42_0
libopencv=4.5.3=py37h595172d_2
libpng=1.6.39=ha978bb4_0
libpq=13.5=h8ce7ee7_1
libprotobuf=3.16.0=hcf210ce_0
libsodium=1.0.18=hbcb3906_1
libsqlite=3.44.2=h92b6c6a_0
libssh2=1.11.0=hd019ec5_0
libtiff=4.3.0=hd146c10_2
libwebp-base=1.3.2=h0dc2134_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=17.0.6=hb6ac08f_0
llvm-tools=12.0.1=hd011deb_2
mako=1.3.0=pyhd8ed1ab_0
markdown=3.5.2=pyhd8ed1ab_0
markupsafe=2.1.1=py37h69ee0a8_1
matplotlib-base=3.5.3=py37h3748cd6_2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
metis=5.1.0=he965462_1007
mpc=1.3.1=h81bd1dd_0
mpfr=4.2.1=h0c69b56_0
mpi=1.0=mpich
mpi4py=3.1.3=py37h483e1c3_2
mpich=4.1.2=hd33e60e_101
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.33=h1d20c9b_6
mysql-libs=8.0.33=hed35180_6
ncurses=6.2=h2e338ed_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.6=hedd7734_0
networkx=2.6.3=pyhd8ed1ab_1
nspr=4.35=hea0b92c_0
nss=3.96=ha05da47_0
numpy=1.21.2=py37h0fa1045_0
numpy-base=1.21.2=py37hbbe2e76_0
olefile=0.47=pyhd8ed1ab_0
openblas=0.3.25=openmp_h6794695_0
opencv=4.5.3=py37hf985489_2
openh264=2.1.1=hfd3ada9_0
openjpeg=2.5.0=h69f46e4_0
openssl=3.2.0=hd75f5a5_1
packaging=23.2=pyhd8ed1ab_0
pandas=1.3.5=py37h5b83a90_0
parso=0.8.2=pyhd8ed1ab_0
patsy=0.5.6=pyhd8ed1ab_0
pcre2=10.42=h0ad2156_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py37h76dc067_0
pip=21.3=pyhd8ed1ab_0
pixman=0.43.0=h73e2aa4_0
pot=0.7.0=py37h570ac47_0
prompt-toolkit=3.0.20=pyha770c72_0
prompt_toolkit=3.0.20=hd8ed1ab_0
protobuf=3.16.0=py37hd8d24ac_0
ptyprocess=0.7.0=pyhd3deb0d_0
py-opencv=4.5.3=py37h4c52d7d_2
pygments=2.10.0=pyhd8ed1ab_0
pygpu=0.7.6=py37h032687b_1003
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.12.3=py37hf985489_8
pyqt-impl=5.12.3=py37hab5ec1f_8
pyqt5-sip=4.19.18=py37h070e122_8
pyqtchart=5.12=py37hab5ec1f_8
pyqtwebengine=5.12.1=py37hab5ec1f_8
pysocks=1.7.1=py37hf985489_5
python=3.7.12=hf3644f1_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0=py37h69ee0a8_4
pyzmq=22.3.0=py37h8f778e5_0
qt=5.12.9=h126340a_4
readline=8.1=h05e3726_0
requests=2.28.2=pyhd8ed1ab_0
scikit-learn=1.0=py37h5a4fc86_1
scipy=1.7.1=py37h4e3cf02_0
seaborn=0.12.2=hd8ed1ab_0
seaborn-base=0.12.2=pyhd8ed1ab_0
setuptools=58.2.0=py37hf985489_0
sigtool=0.1.3=h88f4db0_0
simplegeneric=0.8.1=py_1
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h23a322b_2
statsmodels=0.13.5=py37hacda100_1
suitesparse=5.10.1=h7aff33d_1
sympy=1.10.1=py37hf985489_1
tapi=1100.0.11=h9ce4665_0
tbb=2021.5.0=hbb4e6a2_1
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h3cdfc77_0
tensorflow-base=1.14.0=py37hc8dfbb8_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
tensorly=0.5.1=pyhd3deb0d_0
termcolor=2.2.0=pyhd8ed1ab_0
theano=1.0.5=py37h0582d14_3
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h5dbffcc_1
tornado=6.1=py37h271585c_1
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.0=pyhd8ed1ab_0
typing_extensions=3.10.0.2=pyha770c72_0
urllib3=1.26.18=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.2.3=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
wrapt=1.14.1=py37h994c40b_0
x264=1!161.3030=h0d85af4_1
xz=5.2.5=haf1e3a3_1
yaml=0.2.5=h0d85af4_2
zeromq=4.3.4=he49afe7_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.13=h8a1eda9_5
zstd=1.5.5=h829000d_0
